lib/requirements.txt
# Logging
coloredlogs

# Testing
pytest
pytest-html

# Others
atlassian-python-api